Imposed enemy number limit. See desc.
The enemy spawner script now contains a List. When an enemy is instantiated, it will be added to the list, and will remove itself when destroyed. The spawner will not create more enmeies if the list contains the max number of allowed enemies which can be set from the inspector